+++ src/main/java/org/apiphany/ApiPage.java
@@ -1,8 +1,10 @@
 package org.apiphany;
 
+import java.lang.reflect.Constructor;
 import java.util.List;
 
 import org.apiphany.lang.collections.Lists;
+import org.morphix.reflection.Constructors;
 
 /**
  * API Page class with <code>content</code> list.
@@ -19,6 +21,21 @@ public interface ApiPage<T> {
 	 * @return the content list
 	 */
 	List<T> getContent();
+
+	/**
+	 * Creates an {@link ApiPage} object from an API page derived class that has a constructor with content.
+	 *
+	 * @param <T> content element type
+	 * @param <U> API page type
+	 *
+	 * @param apiPageClass API page class
+	 * @param content content list
+	 * @return API page object
+	 */
+	static <T, U extends ApiPage<T>> U of(final Class<U> apiPageClass, final List<T> content) {
+		Constructor<U> constructor = Constructors.getDeclaredConstructor(apiPageClass, List.class);
+		return Constructors.IgnoreAccess.newInstance(constructor, content);
+	}
 
 	/**
 	 * Returns the first element of the page.