+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added `ApiClient.with` as an alias to `ApiClient.exchangeClient` to create exchange client builder.
 - Added `ExchangeClient.Builder` for managed exchange clients in `ApiClient`.
 - Changed the default exchange client management to not managed when the `ApiClient` is built with direct `ExchangeClient` objects.
+- Added `ApiPage.of` to construct `ApiPage` objects.
 
 ---